% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{add_significance0}
\alias{add_significance0}
\title{Add P-value Significance Symbols}
\usage{
add_significance0(data, p.col = NULL, output.col = NULL)
}
\arguments{
\item{data}{a data frame containing a p-value column.}

\item{p.col}{column name containing p-values.}

\item{output.col}{the output column name to hold the adjusted p-values.}
}
\value{
a data frame
}
\description{
Add p-value significance symbols into a data frame.
This is an wrapper for the function \code{\link[rstatix:add_significance]{rstatix::add_significance()}}.
}
