% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_cat}
\alias{count_cat}
\title{Frequency of categorical variable}
\usage{
count_cat(x, width = 20, collapse = FALSE)
}
\arguments{
\item{x}{Vector or data.frame of categorical variables.}

\item{width}{Maximum width of string.}

\item{collapse}{Boolean to merge categories with identical proportions.}
}
\value{
Data.frame with two columns (f and n).
}
\description{
Formats a data.frame containing categorical variables and calculates the
frequency of each category.
}
