% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{Edge}
\alias{Edge}
\title{Edge (for a network)}
\usage{
Edge(x, y = NULL, digits = 2)
}
\arguments{
\item{x, y}{Symmetrical data.frame with column and row names.}

\item{digits}{Integer for the number of decimals.}
}
\value{
An object edge in the form of a data.frame with three columns: from,
to and weight. The first two columns contain the names of the variable
pairs, and the last, the corresponding values.
}
\description{
Extracts variable pairs with non-zero values as edges.
}
\examples{
x <- sapply(seq(5), function(j) runif(5))
x[x < 0.5] <- diag(x) <- 0
colnames(x) <- rownames(x) <- paste("Variable", seq(5))
x[lower.tri(x)] <- t(x)[lower.tri(x)]
Edge(x)
Edge(x, x)
}
