% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{mcor_test}
\alias{mcor_test}
\title{Multiple correlation}
\usage{
mcor_test(x, pval = FALSE, method = "spearman", method_adjust = "BH")
}
\arguments{
\item{x}{Data.frame of numerical variables.}

\item{pval}{Boolean to return adjusted p-values rather than coefficients.}

\item{method}{Character for the test method ('pearson', 'kendall', or
spearman').}

\item{method_adjust}{Character for the multiple correction test among
'BH', 'BY', 'bonferroni', 'fdr', 'hochberg', 'holm', 'hommel', 'none'}
}
\value{
Data.frame symmetrical containing correlation test results
(coefficients and adjusted p-value) for each pair of variables.
}
\description{
Tests correlation between several variables. This is an wrapper for the
function \code{\link[stats:cor.test]{stats::cor.test()}}.
}
\examples{
library(magrittr)
x <- runif(20)
x <- lapply(
    c(1, -1),
    function(i) sapply(seq(10), function(j) x * i + runif(10, max = 1))
) \%>\%
    Reduce(cbind, .) \%>\%
    set_colnames(paste("Variable", seq(20)))
mcor_test(x)
mcor_test(x, pval = TRUE, method = "pearson", method_adjust = "bonferroni")
}
