% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{palette_discrete}
\alias{palette_discrete}
\title{Color palette}
\usage{
palette_discrete()
}
\value{
Color vector
}
\description{
This is an wrapper for the function \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}}.
}
