% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{plot_alluvial}
\alias{plot_alluvial}
\title{Alluvial diagram}
\usage{
plot_alluvial(
  x,
  width_label = 20,
  colour = palette_discrete(),
  cex = 1,
  cex_axis = 17 * cex
)
}
\arguments{
\item{x}{Data.frame of multiple categorical variables.}

\item{width_label}{Integer for the maximum length of the labels.}

\item{colour}{Color or vector of colors for the categories.}

\item{cex}{Double for the magnification factor for the text relative to the
default.}

\item{cex_axis}{Double for the magnification factor for the axis labels
relative to the default.}
}
\value{
ggplot object.
}
\description{
Display how the frequency of observations are distributed along multiple
categorical variables. The width of the links and the height of the bars
are proportional to the the number of cases. Alluvial diagrams are a good
alternative to multiple pie charts.
}
\examples{
library(ggalluvial)
library(magrittr)
library(RColorBrewer)

x <- lapply(seq(3), function(x) {
    runif(100, 1, 3) \%>\%
        round() \%>\%
        letters[.]
}) \%>\%
    as.data.frame() \%>\%
    set_colnames(paste("variable", LETTERS[seq(3)]))
x[x == "a"] <- NA
x[, 3][is.na(x[, 3])] <- "a"

# Default parameters
plot_alluvial(x)

# Advanced parameters
plot_alluvial(
    x,
    width_label = 5,
    colour = rev(brewer.pal(3, "Reds")),
    cex = 1.5
)
}
