% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Barplot}
\usage{
plot_bar(
  x = NULL,
  title = NULL,
  width_title = 20,
  colour = c("blue", "gray", "#cd5b45"),
  color_title = "black",
  cex = 1,
  cex_main = cex * 30,
  digits = 0,
  n_max = 100,
  ratio = 5,
  threshold = 2,
  hjust_title = 0
)
}
\arguments{
\item{x}{Vector of numerical values visualized on the plot.}

\item{title}{Character for the title.}

\item{width_title}{Integer for the maximum length of the title.}

\item{colour}{Color or vector of colors for the gradient of the bars.}

\item{color_title}{Color for the title.}

\item{cex}{Double for the magnification factor for the text relative to the
default.}

\item{cex_main}{Double for the magnification factor for the subtitles
relative to the default.}

\item{digits}{Integer for the number of decimals.}

\item{n_max}{Integer for the maximum number of bars to show
(prioritizing those with the largest value)}

\item{ratio}{Double for the width scale}

\item{threshold}{Double for the minimal percentage value before being
hidden on the plot.}

\item{hjust_title}{Double for the horizontal justification of the title
(in [0, 1]).}
}
\description{
Display each numerical value separately using a barplot
}
\examples{
library(magrittr)

# Default parameters
x <- runif(10, 1, 10) \%>\%
    set_names(paste("Sample", LETTERS[seq(10)]))
plot_bar(x)

# Advanced parameters
plot_bar(
    x = x,
    title = "Some numerical variable",
    width_title = 30,
    colour = c("yellow", "gray", "red"),
    color_title = "blue",
    cex = 1.2,
    digits = 1,
    n_max = 5,
    ratio = 15,
    hjust_title = 1
)
}
