% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mcor.R
\name{plot_mcor}
\alias{plot_mcor}
\title{Correlation matrix}
\usage{
plot_mcor(
  x,
  colour = brewer.pal(n = 8, name = "RdBu"),
  cex = 1,
  method = "spearman",
  method_adjust = "BH",
  mat = NULL,
  p_mat = NULL
)
}
\arguments{
\item{x}{Data.frame of double variables (with column names).}

\item{colour}{Color or vector of colors for the gradient of the bars.}

\item{cex}{Double for the magnification factor for the text relative to the
default.}

\item{method}{Character for the test method ('pearson' or 'spearman').}

\item{method_adjust}{Character for the multiple correction test among
'BH', 'BY', 'bonferroni', 'fdr', 'hochberg', 'holm', 'hommel', 'none'}
}
\value{
NULL (launch a basic plot)
}
\description{
Display a heatmap of correlation coefficients (with p-values between several
numerical variables). Each cell indicates the correlation between two
variables. The darker the color, the stronger the correlation. Blue
indicates a positive correlation, red a negative one. Non-significant
adjusted p-values are crossed out. This is an wrapper for the function
\code{\link[corrplot:corrplot]{corrplot::corrplot()}}.
}
\examples{
library(magrittr)
library(RColorBrewer)

x <- sapply(seq(20), function(x) runif(20)) \%>\%
    set_colnames(paste("Variable", seq(20)))

# Default parameters
plot_mcor(x)

# Advanced parameters
plot_mcor(
    x,
    colour = c("black", brewer.pal(n = 6, name = "RdBu"), 1),
    method = "spearman",
    method_adjust = "none",
    cex = 0.8
)
}
