% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{plot_network_dyn}
\alias{plot_network_dyn}
\title{Network plot (dynamic)}
\usage{
plot_network_dyn(
  x = NULL,
  title = "",
  cex = 1,
  cex_main = 14 * cex,
  cex_node = 3 * cex,
  cex_edge = 2 * cex,
  color = c("#eee685", "#686868"),
  shape = "dot",
  dashed = TRUE,
  node = NULL,
  edge = NULL
)
}
\arguments{
\item{x}{Vector or data.frame of numerical values visualized on the plot.}

\item{title}{Character for the title.}

\item{cex}{Double for the magnification factor for the text relative to the
default.}

\item{cex_main}{Double for the magnification factor for the subtitles
relative to the default.}

\item{cex_node}{Double for the magnification factor for the node width
relative to the default.}

\item{cex_edge}{Double for the magnification factor for the edge width
relative to the default.}

\item{color}{Color vector of length 2 corresponding respectively to
background and node label.}

\item{shape}{Character for node shape (among: 'circle', 'square' or 'none').}

\item{dashed}{Boolean for dashed edges.}

\item{node}{Node object.}

\item{edge}{Edge object.}
}
\value{
visNetwork object
}
\description{
Plot the connection between variables (dynamic plot). The higher the number
of connections, the larger the node.
}
\examples{
x <- sapply(seq(5), function(j) runif(5))
x[x < 0.5] <- diag(x) <- 0
colnames(x) <- rownames(x) <- paste("Variable", seq(5))
x[lower.tri(x)] <- t(x)[lower.tri(x)]
plot_network(x)
e <- Edge(x)
plot_network_dyn(
    title = "Gimme a network",
    cex = 1.5,
    color = c("white", "gray"),
    shape = "square",
    dashed = FALSE,
    edge = e,
    node = Node(e)
)
}
