% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Radar plot}
\usage{
plot_radar(
  x,
  colour = get_colors(),
  cex = 1,
  n_max = NULL,
  add_percent = FALSE,
  digits = 0,
  n_interval = 2,
  alpha = 0.25
)
}
\arguments{
\item{x}{either a integer vector (with named elements) or a data frame
containing variables in columns and samples in rows (with named columns
and rows).}

\item{colour}{Color or vector of colors for the samples.}

\item{cex}{Double for the size of the labels.}

\item{n_max}{Integer for the maximum value of the axis.}

\item{add_percent}{Boolean to include percentage values in the axis labels.}

\item{digits}{Integer for the number of digits to be displayed.}

\item{n_interval}{Integer for the number of intervals for the axis.}

\item{alpha}{Double for the transparency.}
}
\value{
NULL (launch a basic plot)
}
\description{
Display multivariate data in the form of a two-dimensional chart of three
or more quantitative variables represented on axes starting from the same
point. The relative position and angle of the axes is typically
uninformative. The radar chart is also known as web chart, spider chart,
star chart, cobweb chart, irregular polygon, polar chart, or kiviat diagram.
}
\examples{
library(magrittr)

df <- sapply(
    seq(2),
    function(x) {
        runif(10, 1, 100) \%>\%
            set_names(paste("Variable", letters[seq(10)]))
    }
) \%>\% set_colnames(paste("Sample", LETTERS[seq(2)]))

# Default parameters
plot_radar(df[, 1])

# Advanced parameters
plot_radar(
    df,
    colour = c("orange", "forestgreen"),
    cex = 0.9,
    n_max = 100,
    digits = 1,
    n_interval = 4,
    alpha = 0.5
)

}
