% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{print_mean_test}
\alias{print_mean_test}
\title{Print the result of a mean comparison test}
\usage{
print_mean_test(x, digits = 1, digits_p = 2)
}
\arguments{
\item{x}{Mean comparison test object among \code{anova_test}, \code{kruskal_test} or
\code{wilcox_test}.}

\item{digits}{Integer for the number of decimal places of the test value.}

\item{digits_p}{Integer for the number of decimal places of the p-value.}
}
\value{
String
}
\description{
Print the result of a mean comparison test
}
\examples{
library(rstatix)
data("ToothGrowth")
df <- ToothGrowth
res <- anova_test(df, len ~ dose)
print_mean_test(res)
res <- kruskal_test(df, len ~ dose)
print_mean_test(res)
res <- wilcox_test(df, len ~ supp)
print_mean_test(res)
}
