% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{print_median}
\alias{print_median}
\title{Print the median and standard deviation}
\usage{
print_median(x, digits = 1, width = 10)
}
\arguments{
\item{x}{Vector or integers.}

\item{digits}{Integer for the number of decimals.}

\item{width}{Maximum width of string.}
}
\value{
String
}
\description{
Print the median and standard deviation
}
\examples{
print_median(runif(10))
}
