% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_pretty}
\alias{str_pretty}
\title{Truncate a string to maximum width}
\usage{
str_pretty(x, width = 20)
}
\arguments{
\item{x}{String}

\item{width}{Maximum width of string.}
}
\value{
String
}
\description{
Truncate a string to maximum width while ensuring that whole words are
retained
}
\examples{
str_pretty("Hi there, I'm a sentence to format.")
}
