% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_trunc0}
\alias{str_trunc0}
\title{Truncate a string to maximum number of words.}
\usage{
str_trunc0(x, n = 5, sep = " ")
}
\arguments{
\item{x}{String}

\item{n}{Maximum number of words.}

\item{sep}{Character to separate the terms.}
}
\value{
String
}
\description{
Truncate a string to maximum number of words.
}
