% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_trunc1}
\alias{str_trunc1}
\title{Truncate a string to maximum width}
\usage{
str_trunc1(x, width = 20, sep = " ")
}
\arguments{
\item{x}{String}

\item{width}{Maximum width of string.}

\item{sep}{Character to separate the terms.}
}
\value{
String
}
\description{
Truncate a string to maximum width while ensuring that whole words are
retained
}
\examples{
str_pretty("Hi there, I'm a sentence to format.")
}
