% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_title}
\alias{to_title}
\title{Capitalize  only the first letter}
\usage{
to_title(x)
}
\arguments{
\item{x}{String}
}
\value{
String
}
\description{
Capitalize  only the first letter
}
\examples{
to_title("hi there, I'm a sentence to format.")
}
